task: remove redundant two first examples
