+++ lectures/week3/lib/example_3.dart
@@ -1,74 +1,158 @@
 import 'package:flutter/material.dart';
-import 'package:week3/util/build_aware_widget.dart';
 
-class Example3 extends StatelessWidget {
+class Example3 extends StatefulWidget {
   const Example3({super.key});
 
   @override
+  State<Example3> createState() => _Example3State();
+}
+
+class _Example3State extends State<Example3> {
+  String _arenaWinner = 'Tap to see who wins!';
+
+  @override
   Widget build(BuildContext context) {
-    return DefaultTabController(
-      length: 2,
-      child: Scaffold(
-        appBar: AppBar(
-          leading: const SizedBox.shrink(),
-          title: const Text('SingleChildScrollView vs ListView'),
-          bottom: const TabBar(
-            tabs: [
-              Tab(text: 'SingleChildScrollView'),
-              Tab(text: 'ListView'),
-            ],
-          ),
-        ),
-        body: const TabBarView(
-          children: [_SingleChildScrollViewScreen(), _ListViewScreen()],
-        ),
+    return Scaffold(
+      appBar: AppBar(
+        leading: const SizedBox.shrink(),
+        title: const Text('GestureDetector & Gesture Arena'),
       ),
-    );
-  }
-}
-
-class _SingleChildScrollViewScreen extends StatelessWidget {
-  const _SingleChildScrollViewScreen();
-
-  @override
-  Widget build(BuildContext context) {
-    return Stack(
-      children: [
-        SingleChildScrollView(
-          child: Column(
-            children: [
-              for (var i = 0; i < 1000; i++)
-                BuildAwareWidget(
-                  index: i,
-                  parent: 'SingleChildScrollView',
-                  child: Container(
-                    height: 80,
-                    color: Colors.primaries[i % Colors.primaries.length],
-                  ),
-                ),
-            ],
+      body: Column(
+        children: [
+          // Persistent Status Headers
+          Container(
+            padding: const EdgeInsets.all(16),
+            margin: const EdgeInsets.all(16),
+            decoration: BoxDecoration(
+              color: Colors.amber.shade100,
+              borderRadius: BorderRadius.circular(8),
+            ),
+            child: Text(
+              'Arena Winner: $_arenaWinner',
+              style: const TextStyle(fontSize: 16, fontWeight: FontWeight.bold),
+              textAlign: TextAlign.center,
+            ),
           ),
-        ),
-      ],
-    );
-  }
-}
+          Expanded(
+            child: SingleChildScrollView(
+              child: Padding(
+                padding: const EdgeInsets.symmetric(horizontal: 16),
+                child: Column(
+                  crossAxisAlignment: CrossAxisAlignment.stretch,
+                  children: [
+                    const Text(
+                      '1. Nested Gestures',
+                      style: TextStyle(
+                        fontSize: 14,
+                        fontWeight: FontWeight.w600,
+                      ),
+                    ),
+                    const SizedBox(height: 8),
+                    GestureDetector(
+                      onTap: () => setState(() => _arenaWinner = '🟦 Parent'),
+                      child: Container(
+                        padding: const EdgeInsets.all(40),
+                        decoration: BoxDecoration(
+                          color: Colors.blue,
+                          borderRadius: BorderRadius.circular(12),
+                        ),
+                        child: IgnorePointer(
+                          // ignoring: false,
+                          child: GestureDetector(
+                            onTap: () =>
+                                setState(() => _arenaWinner = '🟩 Child'),
+                            child: Container(
+                              height: 100,
+                              decoration: BoxDecoration(
+                                color: Colors.green,
+                                borderRadius: BorderRadius.circular(8),
+                              ),
+                              child: const Center(
+                                child: Text(
+                                  'Child',
+                                  style: TextStyle(
+                                    color: Colors.white,
+                                    fontWeight: FontWeight.bold,
+                                    fontSize: 16,
+                                  ),
+                                ),
+                              ),
+                            ),
+                          ),
+                        ),
+                      ),
+                    ),
+                    const SizedBox(height: 16),
 
-class _ListViewScreen extends StatelessWidget {
-  const _ListViewScreen();
+                    const Text(
+                      '2. Tap vs DoubleTap',
+                      style: TextStyle(
+                        fontSize: 14,
+                        fontWeight: FontWeight.w600,
+                      ),
+                    ),
+                    const SizedBox(height: 8),
+                    GestureDetector(
+                      onTap: () => setState(() => _arenaWinner = 'Single Tap'),
+                      onDoubleTap: () =>
+                          setState(() => _arenaWinner = 'Double Tap wins!'),
+                      child: Container(
+                        height: 100,
+                        decoration: BoxDecoration(
+                          color: Colors.purple,
+                          borderRadius: BorderRadius.circular(12),
+                        ),
+                        child: const Center(
+                          child: Text(
+                            'Try single vs double tap',
+                            style: TextStyle(
+                              color: Colors.white,
+                              fontWeight: FontWeight.bold,
+                            ),
+                          ),
+                        ),
+                      ),
+                    ),
+                    const SizedBox(height: 16),
 
-  @override
-  Widget build(BuildContext context) {
-    return ListView.builder(
-      itemCount: 1000,
-      itemBuilder: (c, i) => BuildAwareWidget(
-        key: ValueKey('lv-$i'),
-        index: i,
-        parent: 'ListView.builder',
-        child: Container(
-          height: 80,
-          color: Colors.primaries[i % Colors.primaries.length],
-        ),
+                    const Text(
+                      '3. Pan vs LongPress',
+                      style: TextStyle(
+                        fontSize: 14,
+                        fontWeight: FontWeight.w600,
+                      ),
+                    ),
+                    const SizedBox(height: 8),
+                    GestureDetector(
+                      onPanStart: (details) =>
+                          setState(() => _arenaWinner = 'Pan'),
+                      onLongPress: () =>
+                          setState(() => _arenaWinner = 'LongPress'),
+                      child: Container(
+                        height: 100,
+                        decoration: BoxDecoration(
+                          color: Colors.deepOrange,
+                          borderRadius: BorderRadius.circular(12),
+                        ),
+                        child: const Center(
+                          child: Text(
+                            'Hold still or Drag?',
+                            style: TextStyle(
+                              color: Colors.white,
+                              fontWeight: FontWeight.bold,
+                            ),
+                          ),
+                        ),
+                      ),
+                    ),
+
+                    const SizedBox(height: 40),
+                  ],
+                ),
+              ),
+            ),
+          ),
+        ],
       ),
     );
   }
